requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
discord.py==2.4.0
fonttools==4.53.1
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
kiwisolver==1.4.7
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
webargs==8.6.0
Werkzeug==3.0.4
yarl==1.11.1
